Adds forge fmt change

+++ src/libraries/AddressArray.sol
@@ -6,11 +6,10 @@ pragma solidity ^0.8.10;
  * @title Array library
  * @author shAave
  * @dev Implements the logic for manipulating arrays.
-*/
+ */
 library AddressArray {
-
     function removeAddress(address[] storage _array, address _address) internal {
-        for (uint i; i<_array.length; i++) {
+        for (uint256 i; i < _array.length; i++) {
             if (_array[i] == _address) {
                 _array[i] = _array[_array.length - 1];
                 _array.pop();
@@ -20,12 +19,11 @@ library AddressArray {
     }
 
     function includes(address[] memory _array, address _address) internal pure returns (bool) {
-        for (uint i; i<_array.length; i++) {
+        for (uint256 i; i < _array.length; i++) {
             if (_array[i] == _address) {
                 return true;
             }
         }
         return false;
     }
-    
 }
