+++ test/test_child/Child.sell.t.sol
@@ -12,19 +12,19 @@ import ".././mocks/MockUniswap.t.sol";
 import "./common/TestSetup.t.sol";
 import "../common/constants.t.sol";
 
-
 contract SellAllTest is ShaaveChildHelper {
-
     // Contracts
     Child testShaaveChild;
 
     // Events
-    event SwapSuccess(address user, address tokenInAddress, uint tokenInAmount, address tokenOutAddress, uint tokenOutAmount);
+    event SwapSuccess(
+        address user, address tokenInAddress, uint256 tokenInAmount, address tokenOutAddress, uint256 tokenOutAmount
+    );
 
-    
     function setUp() public {
         // Instantiate Child
-        testShaaveChild = new Child(address(this), BASE_TOKEN, IERC20Metadata(BASE_TOKEN).decimals(), getShaaveLTV(BASE_TOKEN));
+        testShaaveChild =
+            new Child(address(this), BASE_TOKEN, IERC20Metadata(BASE_TOKEN).decimals(), getShaaveLTV(BASE_TOKEN));
 
         // Add short position, so we can sell
         deal(BASE_TOKEN, address(testShaaveChild), TEST_COLLATERAL_AMOUNT);
@@ -46,18 +46,25 @@ contract SellAllTest is ShaaveChildHelper {
         Child.PositionData[] memory preAccountingData = testShaaveChild.getAccountingData();
 
         /// @dev Expectations
-        uint baseTokenConversion = 10 ** (18 - IERC20Metadata(BASE_TOKEN).decimals());
-        (uint amountIn, uint amountOut) = swapToShortToken(SHORT_TOKEN, BASE_TOKEN, preAccountingData[0].shortTokenAmountsSwapped[0], preAccountingData[0].backingBaseAmount, baseTokenConversion);
+        uint256 baseTokenConversion = 10 ** (18 - IERC20Metadata(BASE_TOKEN).decimals());
+        (uint256 amountIn, uint256 amountOut) = swapToShortToken(
+            SHORT_TOKEN,
+            BASE_TOKEN,
+            preAccountingData[0].shortTokenAmountsSwapped[0],
+            preAccountingData[0].backingBaseAmount,
+            baseTokenConversion
+        );
         vm.expectEmit(true, true, true, true, address(testShaaveChild));
         emit SwapSuccess(address(this), BASE_TOKEN, amountIn, SHORT_TOKEN, amountOut);
 
         /// @dev Act
-        vm.warp(block.timestamp + 120);    // Trick Aave into thinking it's not a flash loan ;)
+        vm.warp(block.timestamp + 120); // Trick Aave into thinking it's not a flash loan ;)
         assert(testShaaveChild.reducePosition(SHORT_TOKEN, 100, true));
 
         /// @dev Post-action data extraction
         Child.PositionData[] memory postAccountingData = testShaaveChild.getAccountingData();
-        (uint aTokenBalance, uint debtTokenBalance, uint baseTokenBalance, uint userBaseBalance) = getTokenData(address(testShaaveChild), BASE_TOKEN);
+        (uint256 aTokenBalance, uint256 debtTokenBalance, uint256 baseTokenBalance, uint256 userBaseBalance) =
+            getTokenData(address(testShaaveChild), BASE_TOKEN);
 
         /// @dev Assertions
         // Length
@@ -65,7 +72,7 @@ contract SellAllTest is ShaaveChildHelper {
         assertEq(postAccountingData[0].baseAmountsReceived.length, 1);
         assertEq(postAccountingData[0].collateralAmounts.length, 1);
         assertEq(postAccountingData[0].baseAmountsSwapped.length, 1);
-        assertEq(postAccountingData[0].shortTokenAmountsReceived.length, 1); 
+        assertEq(postAccountingData[0].shortTokenAmountsReceived.length, 1);
         assertEq(postAccountingData[0].shortTokenAmountsSwapped.length, 1);
         // Values
         // Esnure this data updated
@@ -79,10 +86,10 @@ contract SellAllTest is ShaaveChildHelper {
         assertEq(postAccountingData[0].shortTokenAddress, preAccountingData[0].shortTokenAddress);
 
         // Enure correct resulting token balances
-        uint baseTolerance = 10000;  // USDC Units: 6 decimals
-        uint debtTolerance = 1000;
-        int baseTokenDiff = int(userBaseBalance) - int(TEST_COLLATERAL_AMOUNT);
-        uint baseTokenDiffAbs = baseTokenDiff < 0 ? uint(-baseTokenDiff) : uint(baseTokenDiff);
+        uint256 baseTolerance = 10000; // USDC Units: 6 decimals
+        uint256 debtTolerance = 1000;
+        int256 baseTokenDiff = int256(userBaseBalance) - int256(TEST_COLLATERAL_AMOUNT);
+        uint256 baseTokenDiffAbs = baseTokenDiff < 0 ? uint256(-baseTokenDiff) : uint256(baseTokenDiff);
         assert(debtTokenBalance < debtTolerance);
         assert(aTokenBalance <= baseTolerance);
         assertEq(baseTokenBalance, 0);
@@ -95,20 +102,27 @@ contract SellAllTest is ShaaveChildHelper {
 
         /// @dev Expectations
         vm.expectEmit(true, true, true, true, address(testShaaveChild));
-        emit SwapSuccess(address(this), BASE_TOKEN, UNISWAP_AMOUNT_IN_PROFIT, SHORT_TOKEN, preAccountingData[0].shortTokenAmountsSwapped[0]);
+        emit SwapSuccess(
+            address(this),
+            BASE_TOKEN,
+            UNISWAP_AMOUNT_IN_PROFIT,
+            SHORT_TOKEN,
+            preAccountingData[0].shortTokenAmountsSwapped[0]
+            );
 
-        /// @dev Mock Uniswap, such that we can ensure a profit.  
+        /// @dev Mock Uniswap, such that we can ensure a profit.
         deal(SHORT_TOKEN, UNISWAP_SWAP_ROUTER, preAccountingData[0].shortTokenAmountsSwapped[0]);
         bytes memory MockUniswapGainsCode = address(new MockUniswapGains()).code;
         vm.etch(UNISWAP_SWAP_ROUTER, MockUniswapGainsCode);
 
         /// @dev Act
-        vm.warp(block.timestamp + 120);    // Trick Aave into thinking it's not a flash loan ;)
+        vm.warp(block.timestamp + 120); // Trick Aave into thinking it's not a flash loan ;)
         bool success = testShaaveChild.reducePosition(SHORT_TOKEN, 100, true);
 
         /// @dev Post-action data extraction
         Child.PositionData[] memory postAccountingData = testShaaveChild.getAccountingData();
-        (uint aTokenBalance, uint debtTokenBalance, uint baseTokenBalance, uint userBaseBalance) = getTokenData(address(testShaaveChild), BASE_TOKEN);
+        (uint256 aTokenBalance, uint256 debtTokenBalance, uint256 baseTokenBalance, uint256 userBaseBalance) =
+            getTokenData(address(testShaaveChild), BASE_TOKEN);
 
         /// @dev Assertions
         assert(success);
@@ -130,7 +144,6 @@ contract SellAllTest is ShaaveChildHelper {
         assertEq(postAccountingData[0].baseAmountsReceived[0], preAccountingData[0].baseAmountsReceived[0]);
         assertEq(postAccountingData[0].collateralAmounts[0], preAccountingData[0].collateralAmounts[0]);
         assertEq(postAccountingData[0].shortTokenAddress, preAccountingData[0].shortTokenAddress);
-        
 
         // Enure correct resulting token balances
         assertEq(debtTokenBalance, 0);
@@ -142,26 +155,29 @@ contract SellAllTest is ShaaveChildHelper {
     function test_reduecePosition_single_close_out_losses() public {
         /// @dev Pre-action assertions
         Child.PositionData[] memory preAccountingData = testShaaveChild.getAccountingData();
-    
+
         /// @dev Expectations
-        uint neededAmountOut = preAccountingData[0].shortTokenAmountsSwapped[0] / UNISWAP_AMOUNT_OUT_LOSSES_FACTOR;
-        uint borrowAmount = getBorrowAmount(TEST_COLLATERAL_AMOUNT, BASE_TOKEN);
+        uint256 neededAmountOut = preAccountingData[0].shortTokenAmountsSwapped[0] / UNISWAP_AMOUNT_OUT_LOSSES_FACTOR;
+        uint256 borrowAmount = getBorrowAmount(TEST_COLLATERAL_AMOUNT, BASE_TOKEN);
         vm.expectEmit(true, true, true, true, address(testShaaveChild));
-        emit SwapSuccess(address(this), BASE_TOKEN, preAccountingData[0].backingBaseAmount, SHORT_TOKEN, neededAmountOut);
+        emit SwapSuccess(
+            address(this), BASE_TOKEN, preAccountingData[0].backingBaseAmount, SHORT_TOKEN, neededAmountOut
+            );
 
         /// @dev Mock Uniswap, such that we can ensure a loss.
-        
+
         deal(SHORT_TOKEN, UNISWAP_SWAP_ROUTER, neededAmountOut);
         bytes memory MockUniswapLossesCode = address(new MockUniswapLosses()).code;
         vm.etch(UNISWAP_SWAP_ROUTER, MockUniswapLossesCode);
 
         /// @dev Act
-        vm.warp(block.timestamp + 120);    // Trick Aave into thinking it's not a flash loan ;)
+        vm.warp(block.timestamp + 120); // Trick Aave into thinking it's not a flash loan ;)
         bool success = testShaaveChild.reducePosition(SHORT_TOKEN, 100, true);
 
         /// @dev Post-action data extraction
         Child.PositionData[] memory postAccountingData = testShaaveChild.getAccountingData();
-        (uint aTokenBalance, uint debtTokenBalance, uint baseTokenBalance, uint userBaseBalance) = getTokenData(address(testShaaveChild), BASE_TOKEN);
+        (uint256 aTokenBalance, uint256 debtTokenBalance, uint256 baseTokenBalance, uint256 userBaseBalance) =
+            getTokenData(address(testShaaveChild), BASE_TOKEN);
 
         /// @dev Assertions
         assert(success);
@@ -174,17 +190,16 @@ contract SellAllTest is ShaaveChildHelper {
         assertEq(postAccountingData[0].shortTokenAmountsReceived.length, 1);
         // Values
         // Esnure this data updated
-        assertEq(postAccountingData[0].baseAmountsSwapped[0],  preAccountingData[0].backingBaseAmount);
+        assertEq(postAccountingData[0].baseAmountsSwapped[0], preAccountingData[0].backingBaseAmount);
         assert(postAccountingData[0].shortTokenAmountsReceived[0] < preAccountingData[0].shortTokenAmountsSwapped[0]);
         assertEq(postAccountingData[0].backingBaseAmount, 0);
-        
+
         // Esnure this data stayed the same
         assertEq(postAccountingData[0].shortTokenAmountsSwapped[0], preAccountingData[0].shortTokenAmountsSwapped[0]);
         assertEq(postAccountingData[0].baseAmountsReceived[0], preAccountingData[0].baseAmountsReceived[0]);
         assertEq(postAccountingData[0].collateralAmounts[0], preAccountingData[0].collateralAmounts[0]);
         assertEq(postAccountingData[0].shortTokenAddress, preAccountingData[0].shortTokenAddress);
         assertEq(postAccountingData[0].hasDebt, true);
-        
 
         // Enure correct resulting token balances
         assert(debtTokenBalance >= borrowAmount / UNISWAP_AMOUNT_OUT_LOSSES_FACTOR);
@@ -193,8 +208,7 @@ contract SellAllTest is ShaaveChildHelper {
         assert(userBaseBalance < TEST_COLLATERAL_AMOUNT); // There were losses
     }
 
-
-    function testCannot_reduecePosition_amount(uint percentageReduction) public {
+    function testCannot_reduecePosition_amount(uint256 percentageReduction) public {
         vm.assume(percentageReduction > 100);
 
         /// @dev Expectations
@@ -206,16 +220,18 @@ contract SellAllTest is ShaaveChildHelper {
 }
 
 contract SellSomeTest is Test, ShaaveChildHelper {
-
     // Contracts
     Child testShaaveChild;
 
     // Events
-    event SwapSuccess(address user, address tokenInAddress, uint tokenInAmount, address tokenOutAddress, uint tokenOutAmount);
+    event SwapSuccess(
+        address user, address tokenInAddress, uint256 tokenInAmount, address tokenOutAddress, uint256 tokenOutAmount
+    );
 
     function setUp() public {
         // Instantiate Child
-        testShaaveChild = new Child(address(this), BASE_TOKEN, IERC20Metadata(BASE_TOKEN).decimals(), getShaaveLTV(BASE_TOKEN));
+        testShaaveChild =
+            new Child(address(this), BASE_TOKEN, IERC20Metadata(BASE_TOKEN).decimals(), getShaaveLTV(BASE_TOKEN));
 
         // Add short position, so we can sell
         deal(BASE_TOKEN, address(testShaaveChild), TEST_COLLATERAL_AMOUNT);
@@ -231,31 +247,39 @@ contract SellSomeTest is Test, ShaaveChildHelper {
         assertEq(preAccountingData[0].shortTokenAmountsReceived.length, 0);
     }
 
-    function test_reduecePosition_some_single(uint reductionPercentage) public {
+    function test_reduecePosition_some_single(uint256 reductionPercentage) public {
         /// @dev Assumptions
         vm.assume(reductionPercentage > 0 && reductionPercentage <= 100);
 
         /// @dev Pre-action data extraction
         Child.PositionData[] memory preAccountingData = testShaaveChild.getAccountingData();
-        (uint pre_aTokenBalance, uint pre_debtTokenBalance, , ) = getTokenData(address(testShaaveChild), BASE_TOKEN);
+        (uint256 pre_aTokenBalance, uint256 pre_debtTokenBalance,,) = getTokenData(address(testShaaveChild), BASE_TOKEN);
 
         /// @dev Expectations
-        uint positionReduction = (getOutstandingDebt(SHORT_TOKEN, address(testShaaveChild)) * reductionPercentage) / 100;
-        uint initialBackingBaseAmount = preAccountingData[0].backingBaseAmount;
-        (uint amountIn, uint amountOut) = swapToShortToken(SHORT_TOKEN, BASE_TOKEN, positionReduction, initialBackingBaseAmount, testShaaveChild.baseTokenConversion());
-        uint expectedGains = getGains(preAccountingData[0].backingBaseAmount, amountIn, testShaaveChild.baseTokenConversion(), reductionPercentage, address(testShaaveChild));
-        uint expectedWithdrawal = getWithdrawal(address(testShaaveChild), amountOut);
+        uint256 positionReduction =
+            (getOutstandingDebt(SHORT_TOKEN, address(testShaaveChild)) * reductionPercentage) / 100;
+        uint256 initialBackingBaseAmount = preAccountingData[0].backingBaseAmount;
+        (uint256 amountIn, uint256 amountOut) = swapToShortToken(
+            SHORT_TOKEN, BASE_TOKEN, positionReduction, initialBackingBaseAmount, testShaaveChild.baseTokenConversion()
+        );
+        uint256 expectedGains = getGains(
+            preAccountingData[0].backingBaseAmount,
+            amountIn,
+            testShaaveChild.baseTokenConversion(),
+            reductionPercentage,
+            address(testShaaveChild)
+        );
+        uint256 expectedWithdrawal = getWithdrawal(address(testShaaveChild), amountOut);
         vm.expectEmit(true, true, true, true, address(testShaaveChild));
         emit SwapSuccess(address(this), BASE_TOKEN, amountIn, SHORT_TOKEN, amountOut);
 
         /// @dev Act
-        vm.warp(block.timestamp + 120);    // Trick Aave into thinking it's not a flash loan ;)
+        vm.warp(block.timestamp + 120); // Trick Aave into thinking it's not a flash loan ;)
         assert(testShaaveChild.reducePosition(SHORT_TOKEN, reductionPercentage, true));
 
         /// @dev Post-action data extraction
         Child.PositionData[] memory postAccountingData = testShaaveChild.getAccountingData();
-        (uint aTokenBalance, uint debtTokenBalance, , ) = getTokenData(address(testShaaveChild), BASE_TOKEN);
-        
+        (uint256 aTokenBalance, uint256 debtTokenBalance,,) = getTokenData(address(testShaaveChild), BASE_TOKEN);
 
         /// @dev Assertions
         // Length
@@ -263,7 +287,7 @@ contract SellSomeTest is Test, ShaaveChildHelper {
         assertEq(postAccountingData[0].baseAmountsReceived.length, 1);
         assertEq(postAccountingData[0].collateralAmounts.length, 1);
         assertEq(postAccountingData[0].baseAmountsSwapped.length, 1);
-        assertEq(postAccountingData[0].shortTokenAmountsReceived.length, 1); 
+        assertEq(postAccountingData[0].shortTokenAmountsReceived.length, 1);
         assertEq(postAccountingData[0].shortTokenAmountsSwapped.length, 1);
         // Values
         // Esnure this data updated
@@ -276,17 +300,20 @@ contract SellSomeTest is Test, ShaaveChildHelper {
         assertEq(postAccountingData[0].collateralAmounts[0], preAccountingData[0].collateralAmounts[0]);
         assertEq(postAccountingData[0].shortTokenAddress, preAccountingData[0].shortTokenAddress);
 
-        // Enure correct resulting token balances    
-        int debtDiff = int(debtTokenBalance) - int(pre_debtTokenBalance - amountOut); // epectedDebt = pre_debtTokenBalance - amountOut
-        uint debtDiffAbs = debtDiff < 0 ? uint(-debtDiff) : uint(debtDiff);
+        // Enure correct resulting token balances
+        int256 debtDiff = int256(debtTokenBalance) - int256(pre_debtTokenBalance - amountOut); // epectedDebt = pre_debtTokenBalance - amountOut
+        uint256 debtDiffAbs = debtDiff < 0 ? uint256(-debtDiff) : uint256(debtDiff);
 
-        uint expectedATokens = pre_aTokenBalance - expectedWithdrawal / testShaaveChild.baseTokenConversion();
-        int aTokenDiff = int(aTokenBalance) - int(expectedATokens);
-        uint aTokenDiffAbs = aTokenDiff < 0 ? uint(-aTokenDiff) : uint(aTokenDiff);
+        uint256 expectedATokens = pre_aTokenBalance - expectedWithdrawal / testShaaveChild.baseTokenConversion();
+        int256 aTokenDiff = int256(aTokenBalance) - int256(expectedATokens);
+        uint256 aTokenDiffAbs = aTokenDiff < 0 ? uint256(-aTokenDiff) : uint256(aTokenDiff);
 
-        assert(debtDiffAbs <= 10);   // An arbitrary maximum tolerance (0.00001%)
+        assert(debtDiffAbs <= 10); // An arbitrary maximum tolerance (0.00001%)
         assert(aTokenDiffAbs <= 10); // An arbitrary maximum tolerance (0.001%)
         assertEq(IERC20(BASE_TOKEN).balanceOf(address(testShaaveChild)), postAccountingData[0].backingBaseAmount);
-        assertEq(IERC20(BASE_TOKEN).balanceOf(address(this)), (expectedGains + expectedWithdrawal) / testShaaveChild.baseTokenConversion());
+        assertEq(
+            IERC20(BASE_TOKEN).balanceOf(address(this)),
+            (expectedGains + expectedWithdrawal) / testShaaveChild.baseTokenConversion()
+        );
     }
 }
